+++ ironfish/src/rpc/routes/chain/followChain.ts
@@ -5,7 +5,7 @@ import bufio from 'bufio'
 import * as yup from 'yup'
 import { Assert } from '../../../assert'
 import { ChainProcessor } from '../../../chainProcessor'
-import { getBlockSize } from '../../../network/utils/block'
+import { getBlockSize } from '../../../network/utils/serializers'
 import { Block, BlockHeader } from '../../../primitives'
 import { BlockSerde } from '../../../primitives/block'
 import { BlockHashSerdeInstance } from '../../../serde'
